the new summarization algorithm
